[Refactor] Move pipeline event into ExecPrimitive
