+++ be/src/exec/pipeline/adaptive/collect_stats_event.cpp
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#include "exec/pipeline/adaptive/event.h"
+#include "exec/pipeline/adaptive/collect_stats_event.h"
 
 #include <atomic>
 #include <condition_variable>
@@ -20,58 +20,21 @@
 #include <utility>
 
 #include "base/failpoint/fail_point.h"
+#include "base/utility/defer_op.h"
+#include "common/status.h"
 #include "common/thread/priority_thread_pool.hpp"
-#include "compute_env/workgroup/work_group.h"
 #include "exec/pipeline/fragment_context.h"
 #include "exec/pipeline/group_execution/execution_group.h"
 #include "exec/pipeline/pipeline.h"
 #include "exec/pipeline/pipeline_driver.h"
 #include "exec/pipeline/primitives/driver_executor.h"
+#include "exec/pipeline/primitives/event.h"
 #include "exec/pipeline/source_operator.h"
 #include "runtime/current_thread.h"
-#include "runtime/exec_env.h"
+#include "runtime/runtime_state.h"
+#include "runtime/service_contexts.h"
 
 namespace starrocks::pipeline {
-
-// ------------------------------------------------------------------------------------
-// Event
-// ------------------------------------------------------------------------------------
-
-void Event::finish(RuntimeState* state) {
-    if (bool expected_finished = false; !_finished.compare_exchange_strong(expected_finished, true)) {
-        return;
-    }
-
-    for (auto& dependee_entry : _dependees) {
-        auto dependee = dependee_entry.lock();
-        if (dependee == nullptr) {
-            continue;
-        }
-        dependee->finish_dependency(state);
-    }
-}
-
-void Event::finish_dependency(RuntimeState* state) {
-    if (_num_finished_dependencies.fetch_add(1) + 1 == _num_dependencies) {
-        process(state);
-    }
-}
-
-void Event::add_dependency(Event* event) {
-    _num_dependencies++;
-    event->_dependees.emplace_back(shared_from_this());
-}
-
-std::string Event::to_string() const {
-    return std::string("Event{") + "name=" + name() + ",_num_dependencies=" + std::to_string(_num_dependencies) +
-           ",_dependees" + std::to_string(_dependees.size()) +
-           ",_num_finished_dependencies=" + std::to_string(_num_finished_dependencies.load()) +
-           ",_finished=" + std::to_string(_finished.load()) + "}";
-}
-
-// ------------------------------------------------------------------------------------
-// CollectStatsSourceInitializeEvent
-// ------------------------------------------------------------------------------------
 
 class CollectStatsSourceInitializeEvent final : public Event {
 public:
@@ -199,43 +162,8 @@ void CollectStatsSourceInitializeEvent::process(RuntimeState* state) {
     }
 }
 
-// ------------------------------------------------------------------------------------
-// DependsAllEventt
-// ------------------------------------------------------------------------------------
-
-class DependsAllEvent final : public Event {
-public:
-    explicit DependsAllEvent(std::vector<EventPtr> events) : _events(std::move(events)) {}
-
-    ~DependsAllEvent() override = default;
-
-    void process(RuntimeState* state) override { finish(state); }
-
-    std::string name() const override { return "depends_all_event"; }
-
-private:
-    std::vector<EventPtr> _events;
-};
-
-// ------------------------------------------------------------------------------------
-// Event factory methods.
-// ------------------------------------------------------------------------------------
-
-EventPtr Event::create_event() {
-    return std::make_shared<Event>();
-}
-
-EventPtr Event::create_collect_stats_source_initialize_event(DriverExecutor* executor,
-                                                             std::vector<Pipeline*>&& pipelines) {
+EventPtr create_collect_stats_source_initialize_event(DriverExecutor* executor, std::vector<Pipeline*>&& pipelines) {
     return std::make_shared<CollectStatsSourceInitializeEvent>(executor, std::move(pipelines));
-}
-
-EventPtr Event::depends_all(const std::vector<EventPtr>& events) {
-    EventPtr merged_event = std::make_shared<DependsAllEvent>(events);
-    for (const auto& event : events) {
-        merged_event->add_dependency(event.get());
-    }
-    return merged_event;
 }
 
 } // namespace starrocks::pipeline
