+++ be/src/exec/pipeline/adaptive/collect_stats_event.h
@@ -1,0 +1,26 @@
+// Copyright 2021-present StarRocks, Inc. All rights reserved.
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     https://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+#pragma once
+
+#include <vector>
+
+#include "exec/pipeline/pipeline_fwd.h"
+#include "exec/pipeline/primitives/event.h"
+
+namespace starrocks::pipeline {
+
+EventPtr create_collect_stats_source_initialize_event(DriverExecutor* executor, std::vector<Pipeline*>&& pipelines);
+
+} // namespace starrocks::pipeline
