+++ be/src/exec/pipeline/pipeline_builder.cpp
@@ -16,7 +16,6 @@
 
 #include <memory>
 
-#include "adaptive/event.h"
 #include "common/config_exec_flow_fwd.h"
 #include "exec/exec_node.h"
 #include "exec/pipeline/adaptive/collect_stats_context.h"
@@ -33,6 +32,7 @@
 #include "exec/pipeline/noop_sink_operator.h"
 #include "exec/pipeline/pipeline.h"
 #include "exec/pipeline/pipeline_fwd.h"
+#include "exec/pipeline/primitives/event.h"
 #include "exec/pipeline/scan/morsel_queue_factory.h"
 #include "exec/pipeline/scan/scan_operator.h"
 #include "exec/pipeline/spill_process_operator.h"
