+++ be/src/exec/pipeline/pipeline_driver.cpp
@@ -29,16 +29,15 @@
 #include "common/status.h"
 #include "common/statusor.h"
 #include "compute_env/compute_env.h"
+#include "compute_env/spill/common.h"
 #include "compute_env/spill/global_spill_manager.h"
 #include "compute_env/spill/operator_mem_resource_manager.h"
 #include "compute_env/workgroup/work_group.h"
-#include "exec/pipeline/adaptive/event.h"
-#include "exec/pipeline/exchange/exchange_sink_operator.h"
 #include "exec/pipeline/fragment_context.h"
 #include "exec/pipeline/pipeline.h"
+#include "exec/pipeline/primitives/event.h"
 #include "exec/pipeline/primitives/pipeline_metrics.h"
 #include "exec/pipeline/query_context.h"
-#include "exec/pipeline/scan/olap_scan_operator.h"
 #include "exec/pipeline/scan/scan_operator.h"
 #include "exec/pipeline/scan/ticketed_morsel_queue.h"
 #include "exec/pipeline/schedule/timeout_tasks.h"
